+++ src/runtime/src/core/mod.rs
@@ -5,8 +5,6 @@ pub mod iterator;
 pub mod koto;
 pub mod list;
 pub mod map;
-pub mod num2;
-pub mod num4;
 pub mod number;
 pub mod os;
 pub mod range;
@@ -25,8 +23,6 @@ pub struct CoreLib {
     pub list: ValueMap,
     pub map: ValueMap,
     pub os: ValueMap,
-    pub num2: ValueMap,
-    pub num4: ValueMap,
     pub number: ValueMap,
     pub range: ValueMap,
     pub string: ValueMap,
@@ -43,8 +39,6 @@ impl Default for CoreLib {
             list: list::make_module(),
             map: map::make_module(),
             os: os::make_module(),
-            num2: num2::make_module(),
-            num4: num4::make_module(),
             number: number::make_module(),
             range: range::make_module(),
             string: string::make_module(),