+++ src/runtime/src/prelude.rs
@@ -4,7 +4,7 @@
 pub use crate::{
     make_runtime_error, runtime_error, type_error, type_error_with_slice, BinaryOp, CallArgs,
     DataMap, ExternalData, ExternalValue, IntRange, KotoFile, KotoIterator, KotoRead, KotoWrite,
-    MetaKey, MetaMap, MetaMapBuilder, Num2, Num4, RuntimeError, RuntimeResult, UnaryOp, Value,
-    ValueIterator, ValueIteratorOutput, ValueKey, ValueList, ValueMap, ValueNumber, ValueString,
-    ValueTuple, ValueVec, Vm, VmSettings,
+    MetaKey, MetaMap, MetaMapBuilder, RuntimeError, RuntimeResult, UnaryOp, Value, ValueIterator,
+    ValueIteratorOutput, ValueKey, ValueList, ValueMap, ValueNumber, ValueString, ValueTuple,
+    ValueVec, Vm, VmSettings,
 };